Add the utils file and move a few functions into there.
